refactor: 工程化重构所有IPC控制器
