+++ src/main/index.ts
@@ -34,23 +34,23 @@ import { createTray, destroyTray, toggleTray } from './tray';
 import { createSessionMainLogger } from './log/mainLog';
 import { startClipboardUrlWatcher, stopClipboardUrlWatcher } from './clipboard/urlWatcher';
 import { createClipboardUrlState } from './clipboard/clipboardUrlState';
-import { registerClipboardIpcHandlers } from './ipc/clipboard';
-import { registerCaptureIpcHandlers } from './ipc/capture';
-import { registerScreenshotHotkeyIpcHandlers } from './ipc/screenshotHotkey';
-import { registerAppIpcHandlers } from './ipc/app';
-import { registerSystemIpcHandlers } from './ipc/system';
-import { registerUpdaterIpcHandlers } from './ipc/updater';
-import { registerWallpaperIpcHandlers } from './ipc/wallpaper';
-import { registerNetIpcHandlers } from './ipc/net';
-import { registerStoreIpcHandlers } from './ipc/store';
-import { registerLogIpcHandlers } from './ipc/log';
-import { registerMusicIpcHandlers } from './ipc/music';
-import { registerHotkeyIpcHandlers } from './ipc/hotkey';
-import { registerIslandIpcHandlers } from './ipc/island';
-import { registerHideProcessIpcHandlers } from './ipc/hideProcess';
-import { registerThemeIpcHandlers } from './ipc/theme';
-import { registerWindowIpcHandlers } from './ipc/window';
-import { registerMediaIpcHandlers } from './ipc/media';
+import { registerClipboardIpcHandlers } from './ipc/settings/clipboard';
+import { registerCaptureIpcHandlers } from './ipc/window/capture';
+import { registerScreenshotHotkeyIpcHandlers } from './ipc/system/screenshotHotkey';
+import { registerAppIpcHandlers } from './ipc/app/app';
+import { registerSystemIpcHandlers } from './ipc/system/system';
+import { registerUpdaterIpcHandlers } from './ipc/app/updater';
+import { registerWallpaperIpcHandlers } from './ipc/window/wallpaper';
+import { registerNetIpcHandlers } from './ipc/app/net';
+import { registerStoreIpcHandlers } from './ipc/app/store';
+import { registerLogIpcHandlers } from './ipc/app/log';
+import { registerMusicIpcHandlers } from './ipc/media/music';
+import { registerHotkeyIpcHandlers } from './ipc/system/hotkey';
+import { registerIslandIpcHandlers } from './ipc/settings/island';
+import { registerHideProcessIpcHandlers } from './ipc/system/hideProcess';
+import { registerThemeIpcHandlers } from './ipc/settings/theme';
+import { registerWindowIpcHandlers } from './ipc/window/window';
+import { registerMediaIpcHandlers } from './ipc/media/media';
 import { broadcastSettingChange, registerSettingsPreviewHandler } from './utils/broadcast';
 import { registerAppLifecycleHandlers } from './services/appLifecycle';
 import { applyChromiumPerformanceFlags } from './services/chromiumFlags';
