+++ src/main/ipc/app/log.ts
