+++ src/main/ipc/app/store.ts
@@ -28,7 +28,7 @@
 import { ipcMain } from 'electron';
 import { existsSync, readFileSync, writeFileSync } from 'fs';
 import { join } from 'path';
-import { broadcastSettingChange } from '../utils/broadcast';
+import { broadcastSettingChange } from '../../utils/broadcast';
 
 interface RegisterStoreIpcHandlersOptions {
   storeDir: string;
