+++ src/main/ipc/media/music.ts
