+++ src/main/ipc/media/wallpaperVideo.ts
