+++ src/main/ipc/system/hideProcess.ts
