+++ src/main/ipc/system/hotkey.ts
