+++ src/main/ipc/system/screenshotHotkey.ts
