+++ src/main/ipc/window/capture.ts
