+++ src/main/ipc/window/wallpaper.ts
